# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patuscada API",
        "description": "Documentação da API Patuscada em Laravel 8",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cartasbrancas": {
            "get": {
                "tags": [
                    "Cartas Brancas"
                ],
                "summary": "Listar todas as cartas brancas",
                "operationId": "c73beeacae05b3342385c5b8d6f7aa5b",
                "responses": {
                    "200": {
                        "description": "Lista de cartas brancas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "texto": {
                                                "type": "string",
                                                "example": "Uma resposta engraçada"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cartas Brancas"
                ],
                "summary": "Cadastrar nova carta branca",
                "operationId": "02118e081c951dc1ead9982594ecb010",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "texto"
                                ],
                                "properties": {
                                    "texto": {
                                        "type": "string",
                                        "example": "Algo engraçado para jogar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carta branca criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "texto": {
                                            "type": "string",
                                            "example": "Algo engraçado para jogar"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cartasbrancas/{id}": {
            "get": {
                "tags": [
                    "Cartas Brancas"
                ],
                "summary": "Obter uma carta branca por ID",
                "operationId": "33395732a92db0c2443cd3ae9026aa86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da carta branca",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carta branca encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "texto": {
                                            "type": "string",
                                            "example": "Uma resposta engraçada"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Carta branca não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cartas Brancas"
                ],
                "summary": "Atualizar uma carta branca",
                "operationId": "17e9d571170300f5b352e65ad8feb163",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da carta branca a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "texto"
                                ],
                                "properties": {
                                    "texto": {
                                        "type": "string",
                                        "example": "Texto atualizado da carta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carta branca atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "texto": {
                                            "type": "string",
                                            "example": "Texto atualizado da carta"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Carta branca não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cartas Brancas"
                ],
                "summary": "Deletar uma carta branca",
                "operationId": "61390aaec899f6669dd12c66884b0476",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da carta branca a ser deletada",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carta branca deletada com sucesso"
                    },
                    "404": {
                        "description": "Carta branca não encontrada"
                    }
                }
            }
        },
...
    },
...
}